refactor: split attribution active benchmark exposure
Evidence:

- python -m pytest tests\unit\test_attribution_mode_adapter.py tests\unit\test_attribution_engine.py tests\unit\test_attribution_contract.py -q: 34 passed

- python -m ruff check src\app\services\attribution_stateful_inputs.py src\app\services\attribution_active_benchmark_exposure.py scripts\generate_quality_baseline.py: passed

- python -m mypy --config-file mypy.ini src\app\services\attribution_stateful_inputs.py src\app\services\attribution_active_benchmark_exposure.py src\app\services\attribution_mode_adapter.py: passed

- python -m radon mi/cc src\app\services\attribution_stateful_inputs.py src\app\services\attribution_active_benchmark_exposure.py: MI A, no C-or-worse complexity

- python scripts\generate_quality_baseline.py: passed and refreshed baseline evidence

- make check: passed, including 551 unit tests
