+++ src/app/services/attribution_stateful_inputs.py
@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from datetime import date
 from typing import Any, Protocol
 
 from app.contracts.attribution import (
@@ -10,7 +9,11 @@ from app.contracts.attribution import (
     HistoricalAttributionStatefulInput,
     HistoricalAttributionStatelessInput,
 )
-from app.contracts.risk import ReturnPoint, RiskRequestScope
+from app.contracts.risk import RiskRequestScope
+from app.services.attribution_active_benchmark_exposure import (
+    BenchmarkExposureClientProtocol,
+    fetch_active_benchmark_exposure_history,
+)
 from app.services.attribution_exposure_history import (
     fetch_stateful_exposure_history,
 )
@@ -21,11 +24,6 @@ from app.services.attribution_stateful_returns import (
     fetch_stateful_returns_context,
     requires_active_attribution,
 )
-from app.services.benchmark_exposure_history import (
-    BenchmarkExposureHistoryRequest,
-    fetch_benchmark_exposure_history,
-)
-from app.upstream_errors import missing_upstream_data
 
 
 __all__ = [
@@ -39,13 +37,12 @@ __all__ = [
 ]
 
 
-class LotusPerformanceClientProtocol(LotusPerformanceReturnsClientProtocol, Protocol):
-    async def get_benchmark_exposure_context(
-        self,
-        *,
-        request_payload: dict[str, Any],
-        correlation_id: str | None,
-    ) -> dict[str, Any]: ...
+class LotusPerformanceClientProtocol(
+    LotusPerformanceReturnsClientProtocol,
+    BenchmarkExposureClientProtocol,
+    Protocol,
+):
+    pass
 
 
 class LotusCoreClientProtocol(Protocol):
@@ -77,58 +74,11 @@ class _StatefulExposureHistories:
     benchmark_exposure_history: list[ExposurePoint]
 
 
-def _validate_benchmark_exposure_alignment(
-    *,
-    benchmark_returns: list[ReturnPoint],
-    benchmark_exposure_history: list[ExposurePoint],
-) -> None:
-    return_dates = {point.date for point in benchmark_returns}
-    exposure_dates = {point.date for point in benchmark_exposure_history}
-    missing_dates = sorted(return_dates - exposure_dates)
-    if missing_dates:
-        sample = ", ".join(date_value.isoformat() for date_value in missing_dates[:5])
-        raise missing_upstream_data(
-            service="lotus-performance",
-            operation="/integration/benchmarks/exposure-context",
-            message=(
-                "lotus-performance benchmark exposure context missing rows for benchmark "
-                f"return dates: {sample}"
-            ),
-        )
-
-
 def validate_stateful_groupings(grouping_dimensions: list[GroupingDimension]) -> None:
     if "CUSTOM" in grouping_dimensions:
         raise ValueError(
             "stateful historical-attribution does not support grouping_dimension=CUSTOM"
         )
-
-
-async def _fetch_active_benchmark_exposure_history(
-    *,
-    stateful: HistoricalAttributionStatefulInput,
-    performance_client: LotusPerformanceClientProtocol,
-    benchmark_returns: list[ReturnPoint],
-    start_date: date,
-    grouping_dimensions: list[GroupingDimension],
-    correlation_id: str | None,
-) -> list[ExposurePoint]:
-    benchmark_exposure_history = await fetch_benchmark_exposure_history(
-        BenchmarkExposureHistoryRequest(
-            performance_client=performance_client,
-            portfolio_id=stateful.portfolio_id,
-            as_of_date=stateful.as_of_date,
-            start_date=start_date,
-            reporting_currency=stateful.reporting_currency,
-            grouping_dimensions=grouping_dimensions,
-            correlation_id=correlation_id,
-        )
-    )
-    _validate_benchmark_exposure_alignment(
-        benchmark_returns=benchmark_returns,
-        benchmark_exposure_history=benchmark_exposure_history,
-    )
-    return benchmark_exposure_history
 
 
 def build_stateful_stateless_input(
@@ -171,7 +121,7 @@ async def _stateful_exposure_histories(
         correlation_id=correlation_id,
     )
     benchmark_exposure_history = (
-        await _fetch_active_benchmark_exposure_history(
+        await fetch_active_benchmark_exposure_history(
             stateful=stateful,
             performance_client=performance_client,
             benchmark_returns=returns_context.benchmark_returns,
